Contains updated shrink and grow
Can now shrink and grow in any direction, which is set in the inspector. The object you have to shrink or grow has to be the child of an empty game object. The child object has to be positioned relative to the parent object so that the pivot is set correctly to allow for unidirectional scaling. Setting the child object to 0.5 on the transform's x position makes it so that it scales horizontally to the right because the pivot point is set to the left side. Setting the child object to 0.5 on the transform's y position makes it so that it scales vertically upward because the pivot point is set to the bottom.